+++ client/src/layouts/MainLayout.tsx
@@ -4,22 +4,21 @@ import { drawerWidth } from 'src/utils/constants';
 import SidebarLeft from './components/SidebarLeft/SidebarLeft';
 import SidebarRight from './components/SidebarRight/SidebarRight';
 import { ReactNode, Dispatch, SetStateAction } from 'react';
+import useDrawer from 'src/hooks/useDrawer';
 
 type TMainLayoutProps = {
   children: ReactNode[];
-  setIsClosing: Dispatch<SetStateAction<boolean>>;
-  mobileOpen: boolean;
-  setMobileOpen: Dispatch<SetStateAction<boolean>>;
-  handleDrawerToggle: () => void; 
+ 
 }
 
-export default function MainLayout({children, setIsClosing,mobileOpen,setMobileOpen, handleDrawerToggle }:TMainLayoutProps) {
+export default function MainLayout({children }:TMainLayoutProps) {
+  
   
   return (
     <Box sx={{ display: 'flex', flexDirection:'column',color:'#fff', 'header': {boxShadow:'1px solid #fff'} }} >     
-      <TopBar inViewport={true} handleDrawerToggle={handleDrawerToggle}/>     
+      <TopBar inViewport={true}/>     
       <Box sx={{ display: 'flex', backgroundColor: '#000', color: '#fff', zIndex: '1' }}>
-        <SidebarLeft setIsClosing={setIsClosing} mobileOpen={mobileOpen} setMobileOpen={setMobileOpen} />
+        <SidebarLeft />
         {children}
         <SidebarRight drawerWidth={drawerWidth} />
       </Box>
